修改后厨逻辑 修改图片上传

+++ ManageAPI/ManageAPI/Dtos/CancelDishRequest.cs
@@ -4,6 +4,6 @@ namespace ManageAPI.Dtos
     {/// <summary>
      /// 菜品訂單明細 ID
      /// </summary>
-        public int DishOrderDetailsId { get; set; }
+        public long DishOrderDetailsId { get; set; }
     }
 }
